discord-plex-bot/requirements.txt
discord.py>=2.0.0
supabase-py
plexapi
python-dotenv
httpx
asyncpg 